shoe-orientation/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
Pillow==10.4.0
torch>=2.2.0
torchvision>=0.17.0
timm>=1.0.7
numpy>=1.26.0
scikit-learn>=1.4.0
tqdm>=4.66.0
fastapi==0.115.8
uvicorn[standard]==0.30.6
